Added test revealing a regression in Substance.
