+++ test/integration/Persistence.test.js
@@ -1,30 +1,39 @@
 import { test } from 'substance-test'
-import { isString } from 'substance'
+import { DefaultDOMElement as DOM } from 'substance'
 import { TextureWebApp, TextureArchive, checkArchive } from 'substance-texture'
 import { spy, getMountPoint } from '../testHelpers'
+import { applyNOP } from './integrationTestHelpers'
 
-test("Integration: loading and saving the kitchen-sink article", (t) => {
-  t.plan(1)
+test("Persistence: loading and saving the kitchen-sink article", (t) => {
+  t.plan(2)
   let app, archive, manuscriptSession
   let rawArchive
   _setupApp(t)
-  .then((res) => {
-    ({app, archive, manuscriptSession} = res)
-    setCursor(manuscriptSession, 'p-1.content', 0)
-    insertText(manuscriptSession, 'foo')
+  .then(res => {
+    ({app, archive, manuscriptSession } = res)
+    applyNOP(manuscriptSession)
     spy(archive.storage, 'write')
   })
-  .then(() => app._save())
+  .then(() => {
+    return app._save()
+  })
   .then(() => {
     rawArchive = archive.storage.write.args[1]
     return archive.storage.read('kitchen-sink')
   })
   .then(originalRawArchive => {
+    // the XML should have actually not changed
+    let originalManuscriptXML = originalRawArchive.resources['manuscript.xml'].data
+    let newManuscriptXML = rawArchive.resources['manuscript.xml'].data
     // retrieve the rawArchive from the function spy
     rawArchive.resources = Object.assign({}, originalRawArchive.resources, rawArchive.resources)
     let err = checkArchive(TextureArchive, rawArchive)
     let details = err ? err.detail : null
     t.nil(details, 'There should be no error.')
+    // just pick the body and see if this is the same
+    let oldBodyXML = DOM.parseXML(originalManuscriptXML).find('body').getInnerXML()
+    let newBodyXML = DOM.parseXML(newManuscriptXML).find('body').getInnerXML()
+    t.equal(newBodyXML, oldBodyXML)
   })
 })
 
@@ -33,7 +42,6 @@ function _setupApp(t) {
   return new Promise((resolve, reject) => {
     let App = createTestApp(resolve, reject)
     let el = getMountPoint(t)
-    console.log('##### t.sandbox')
     app = App.mount({
       archiveId: 'kitchen-sink',
       storageType: 'vfs'
@@ -63,23 +71,4 @@ function createTestApp(resolve, reject) {
     }
   }
   return App
-}
-
-function setCursor(editorSession, path, pos) {
-  if (isString(path)) path = path.split('.')
-  let doc = editorSession.getDocument()
-  let node = doc.get(path[0])
-  editorSession.setSelection({
-    type: 'property',
-    path,
-    startOffset: pos,
-    endOffset: pos,
-    surfaceId: node.parentNode.id
-  })
-}
-
-function insertText(editorSession, text) {
-  editorSession.transaction(tx => {
-    tx.insertText(text)
-  })
 }
